-3.0.6-test2, before switching to their physical params

+++ end2end/model.py
@@ -229,9 +229,9 @@ class AchromaticEdofFourier(nn.Module):
 
         # FIXME: is this best way of ading dims?
         squared_sum = torch.unsqueeze(torch.unsqueeze(grid_x ** 2 + grid_y ** 2, 0), 0)
-        curvature = torch.sqrt(squared_sum + depth_map ** 2) # [m, c, H, W]
-        wave_nos = 2. * np.pi / self.wave_lengths # [c,]
-        wave_nos = wave_nos.reshape([1, -1, 1, 1])
+        curvature = torch.sqrt(squared_sum + depth_map ** 2)  # [m, c, H, W]
+        wave_nos = 2. * np.pi / self.wave_lengths  # [c,]
+        wave_nos = wave_nos.reshape([1, -1, 1, 1]).type(torch.float32)
         input_field = torch.exp(torch.complex(torch.zeros_like(curvature), wave_nos * curvature))
 
         field = self.heightMapElement(input_field)
@@ -279,7 +279,7 @@ class AchromaticEdofFourier(nn.Module):
         squared_sum = torch.unsqueeze(torch.unsqueeze(grid_x ** 2 + grid_y ** 2, 0), 0)
         curvature = torch.sqrt(squared_sum + depth_map ** 2)
         wave_nos = 2. * np.pi / self.wave_lengths  # [c,]
-        wave_nos = wave_nos.reshape([1, -1, 1, 1])
+        wave_nos = wave_nos.reshape([1, -1, 1, 1]).type(torch.float32)
         input_field = torch.exp(torch.complex(torch.zeros_like(curvature), wave_nos * curvature))
 
         field = self.heightMapElement(input_field)
